requirements.txt
certifi==2019.11.28
chardet==3.0.4
crypto==1.4.1
idna==2.8
Naked==0.1.31
Pillow==7.0.0
pycryptodome==3.9.6
pytesseract==0.3.2
PyYAML==5.3
requests==2.7.0
shellescape==3.8.1
urllib3==1.25.8
python-telegram-bot==12.6.1
